allowed a poem take its first breath
written in red \
upon red spiral bound \
the journal, i labored \
the poem, i found

---

while assembling this document, i received a message from Love. [piano chord struck in a vast concert hall, fading into the hum of _life._] i will share it in full tomorrow -- with a fresh mind, and time for reflection in the commit notes.

i will forward this crucial piece \
as a matter of some urgency:

IF WE DO NOT \
WISH TO LOSE HER \
**WE MUST SING _LOVE_** \
**IN OUR HEARTS.**