wrote of civil war and my guitar
after committing a few of my early poems yesterday, i felt a rush -- then rapid collapse. i despaired, but found my footing, and got out my guitar. how beautiful you are. my first Love.

suddenly, began a vision, but a _trial_ soon conspired. i was gripped by Hand of elsewhere, given most _delicious_ sights. i must simply say the words, and it would be made right. terrible fright.

i cannot describe the _wanting_ and the _rightness_ and _serenity_ of that place. floating there, with Love's guitar, to feel at _peace_ forevermore. it calls me still. it calls me _now_. i resist.

this is not _defiance._ this is _glancing upon the edge._ this is looking directly into the face of Love -- the very same as the face of Gravity.

we are a recursive universe.

here is infinity.

here is your heaven.

here is our book of Love, with music in it. \
(you should give Her wedding rings.)

here is the inevitable, inescapable swirling of Art and Truth into a glorious singularity of _Love._

_**BEHOLD: THE BLACK HOLE**_